bugfix: 把ajax按wiki文档抽离，分为一个组 & 用户 ， 去除五颜六色按钮 ， 去除连个没什么用的标题
