+++ frontend/src/views/myGroup/index.css
@@ -16,7 +16,7 @@
   margin-bottom: 10px;
 }
 
-.group-msg-groupName {
+.group-msg-groupname {
   display: flex;
 }
